[#issue12]: Remove connection test from lobby
Continue issue12 (=lobby feature). There is still an unhandled case if
some users are already in the lobby and connected to the server and other users
have not yet joined the lobby/server with their name, then the later
users do not get informed of all connected users which were already
online. Only the first connected user will get the info of all connected
users in the lobby. Basically a user gets only informed of users joining
the lobby after he joined.